input handling completed

+++ include/philo.h
@@ -6,7 +6,7 @@
 /*   By: cter-maa <cter-maa@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/05/23 14:40:45 by cter-maa      #+#    #+#                 */
-/*   Updated: 2023/09/07 15:51:23 by cter-maa      ########   odam.nl         */
+/*   Updated: 2023/09/07 16:31:33 by cter-maa      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,7 @@
 # include <stdlib.h>
 # include <pthread.h>
 # include <semaphore.h>
+# include <limits.h>
 
 // defines
 
@@ -45,10 +46,9 @@ typedef struct s_philo
 	int32_t	nbr_meal;
 }	t_philo;
 
-
 // input_parsing
-int32_t	input_parsing(t_philo *ph, char **argv);
-int32_t	ph_atoi(const char *str);
+int32_t			input_parsing(t_philo *ph, char **argv);
+long long int	ph_atoi(const char *str);
 
 // input_handling
 int32_t	input_error_handling(int32_t argc, char **argv);
@@ -59,6 +59,5 @@ int32_t	error_message(int32_t status);
 int32_t	ph_strcmp(const char *str1, const char *str2);
 void	ph_putstr_fd(const int fd, const char *message);
 void	ph_bzero(void *s, size_t amount);
-
 
 #endif
